Начинаю делать presenter
