+++ src/presenter/points.js
@@ -1,0 +1,16 @@
+
+export default class points {
+  constructor() {
+  }
+
+}
+
+
+_handleSortTypeChange() {
+}
+
+_renderSort() {
+}
+
+_renderTask() {
+}
