+++ src/view/mainTripFilter.js
@@ -1,10 +1,7 @@
-import { createElement } from "./util.js";
+import Abstract from "./abstract.js";
 
 
-export default class TripFilterView {
-  constructor() {
-    this._element = null;
-  }
+export default class TripFilterView extends Abstract {
 
   createMainTripFilterTemplate() {
     return `<form class="trip-filters" action="#" method="get">
@@ -29,17 +26,5 @@ export default class TripFilterView {
 
   getTemplate() {
     return this.createMainTripFilterTemplate()
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
