+++ src/view/no-Points.js
@@ -1,28 +1,9 @@
-import { createElement } from "./util.js";
 
+import Abstract from "./abstract.js";
 
-export default class NoPoints {
-  constructor() {
-    this._element = null;
-  }
+export default class NoPoints extends Abstract {
 
   createNoPointTemplate() {
     return `<p class="trip-events__msg">Click New Event to create your first point</p>`;
-  }
-
-  getTemplate() {
-    return this.createNoPointTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
