+++ src/view/pointEditor.js
@@ -1,12 +1,14 @@
 
-import { humanizeTaskDueDate, formatedStartEndDate, createElement } from "./util.js";
+import { humanizeTaskDueDate, formatedStartEndDate } from "../utils/dates.js";
 import { getOptions } from "../mock/point.js";
+import Abstract from "./abstract.js";
 
-export default class PointEditView {
+export default class PointEditView extends Abstract {
   constructor(point, points) {
+    super()
     this._point = point;
     this._points = points;
-    this._element = null;
+    this._editClickHandler = this._editClickHandler.bind(this);
   }
 
   createTripEditTemplate(point, points) {
@@ -183,15 +185,13 @@ export default class PointEditView {
     return this.createTripEditTemplate(this._point, this._points);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
+  _editClickHandler(evt) {
+    evt.preventDefault();
+    this._callback.editClick();
   }
 
-  removeElement() {
-    this._element = null;
+  setEditClickHandler(callback) {
+    this._callback.editClick = callback;
+    this.getElement().querySelector(`.event__rollup-btn`).addEventListener(`click`, this._editClickHandler);
   }
 }
